feat: implemented security with JWT in both back & front
Breaking Change:

The implementation of security with JWT in both the back-end and front-end introduces a breaking change. Previously, the system allowed unauthenticated access to certain endpoints, but with the introduction of JWT security, all endpoints now require authentication. This means that clients or users who were previously able to access these endpoints without authentication will now need to provide a valid JWT token in order to access the protected resources.